requirements.txt
influxdb
requests
absl-py
typing
schedule
pytz
pymongo